+++ packages/proteus/src/main/session/Session.ts
@@ -117,7 +117,9 @@ export class Session {
       session._insert_session_state(preKeyMessage.message.session_tag, state);
 
       if (preKeyMessage.prekey_id < PreKey.MAX_PREKEY_ID) {
-        MemoryUtil.zeroize(await prekey_store.load_prekey(preKeyMessage.prekey_id));
+        const prekey = await prekey_store.load_prekey(preKeyMessage.prekey_id);
+        MemoryUtil.zeroize(prekey);
+
         try {
           await prekey_store.delete_prekey(preKeyMessage.prekey_id);
         } catch (error) {
@@ -148,7 +150,7 @@ export class Session {
       );
     }
     throw new ProteusError(
-      `Unable to find PreKey ID "${preKeyMessage.prekey_id}" in PreKey store "${preKeyStore.constructor.name}".`,
+      `Unable to find PreKey with ID "${preKeyMessage.prekey_id}" in PreKey store "${preKeyStore.constructor.name}".`,
       ProteusError.CODE.CASE_101,
     );
   }
@@ -164,7 +166,7 @@ export class Session {
 
       this.session_states[sessionTag.toString()] = {
         idx: this.counter,
-        state: state,
+        state,
         tag: sessionTag,
       };
       this.counter++;
@@ -188,7 +190,7 @@ export class Session {
   private _evict_oldest_session_state(): void {
     const oldest = Object.keys(this.session_states)
       .filter(obj => obj.toString() !== this.session_tag.toString())
-      .reduce((lowest, obj, index) => {
+      .reduce((lowest, obj) => {
         return this.session_states[obj].idx < this.session_states[lowest].idx ? obj.toString() : lowest;
       });
 
@@ -257,8 +259,8 @@ export class Session {
         const plaintext = await state.decrypt(envelope, msg.message);
 
         if (msg.prekey_id !== PreKey.MAX_PREKEY_ID) {
-          // TODO: Zeroize should be tested (and awaited) here!
-          MemoryUtil.zeroize(await prekey_store.load_prekey(msg.prekey_id));
+          const prekey = await prekey_store.load_prekey(msg.prekey_id);
+          MemoryUtil.zeroize(prekey);
           await prekey_store.delete_prekey(msg.prekey_id);
         }